requirements.txt
streamlit
oracledb
bcrypt
python-dotenv